+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [1.0.2] - 2024-12
+
+### Added
+- User command to display authenticated user and other user info by ID
+- Shop command to pull info about a shop by its IP, user ID, and search by name
+- Image command with full support for get, upload, update, and delete
+- Listing Property command with support to retrieve, update, and delete values
+
+### Updated
+- Minor bug fixes and documentation updates
+
 ## [1.0.1] - 2024-12
 
 ### Added